+++ model/src/gateway/event/kind.rs
@@ -128,15 +128,15 @@ impl EventType {
             Self::InviteCreate => Some("INVITE_CREATE"),
             Self::InviteDelete => Some("INVITE_DELETE"),
             Self::MemberAdd => Some("GUILD_MEMBER_ADD"),
+            Self::MemberChunk => Some("GUILD_MEMBERS_CHUNK"),
             Self::MemberRemove => Some("GUILD_MEMBER_REMOVE"),
             Self::MemberUpdate => Some("GUILD_MEMBER_UPDATE"),
-            Self::MemberChunk => Some("GUILD_MEMBERS_CHUNK"),
             Self::MessageCreate => Some("MESSAGE_CREATE"),
             Self::MessageDelete => Some("MESSAGE_DELETE"),
             Self::MessageDeleteBulk => Some("MESSAGE_DELETE_BULK"),
             Self::MessageUpdate => Some("MESSAGE_UPDATE"),
-            Self::PresenceUpdate => Some("PRESENCE_UPDATE"),
             Self::PresencesReplace => Some("PRESENCES_REPLACE"),
+            Self::PresenceUpdate => Some("PRESENCE_UPDATE"),
             Self::ReactionAdd => Some("MESSAGE_REACTION_ADD"),
             Self::ReactionRemove => Some("MESSAGE_REACTION_REMOVE"),
             Self::ReactionRemoveAll => Some("MESSAGE_REACTION_REMOVE_ALL"),
@@ -152,8 +152,8 @@ impl EventType {
             Self::ThreadCreate => Some("THREAD_CREATE"),
             Self::ThreadDelete => Some("THREAD_DELETE"),
             Self::ThreadListSync => Some("THREAD_LIST_SYNC"),
-            Self::ThreadMemberUpdate => Some("THREAD_MEMBER_UPDATE"),
             Self::ThreadMembersUpdate => Some("THREAD_MEMBERS_UPDATE"),
+            Self::ThreadMemberUpdate => Some("THREAD_MEMBER_UPDATE"),
             Self::ThreadUpdate => Some("THREAD_UPDATE"),
             Self::TypingStart => Some("TYPING_START"),
             Self::UnavailableGuild => Some("UNAVAILABLE_GUILD"),
@@ -170,8 +170,8 @@ impl EventType {
             | Self::ShardConnecting
             | Self::ShardDisconnected
             | Self::ShardIdentifying
-            | Self::ShardReconnecting
             | Self::ShardPayload
+            | Self::ShardReconnecting
             | Self::ShardResuming => None,
         }
     }
@@ -337,9 +337,9 @@ mod tests {
         assert_variant(EventType::InviteCreate, "INVITE_CREATE");
         assert_variant(EventType::InviteDelete, "INVITE_DELETE");
         assert_variant(EventType::MemberAdd, "GUILD_MEMBER_ADD");
+        assert_variant(EventType::MemberChunk, "GUILD_MEMBERS_CHUNK");
         assert_variant(EventType::MemberRemove, "GUILD_MEMBER_REMOVE");
         assert_variant(EventType::MemberUpdate, "GUILD_MEMBER_UPDATE");
-        assert_variant(EventType::MemberChunk, "GUILD_MEMBERS_CHUNK");
         assert_variant(EventType::MessageCreate, "MESSAGE_CREATE");
         assert_variant(EventType::MessageDelete, "MESSAGE_DELETE");
         assert_variant(EventType::MessageDeleteBulk, "MESSAGE_DELETE_BULK");
