+++ model/src/guild/audit_log/event_type.rs
@@ -209,19 +209,19 @@ pub enum AuditLogEventType {
     ///
     /// [GuildCommandPermissions]: crate::application::command::permissions::GuildCommandPermissions
     ApplicationCommandPermissionUpdate = 121,
-    /// [`AutoModerationRule`] was created.
+    /// [`AutoModerationRule`] has been created.
     ///
     /// [`AutoModerationRule`]: crate::guild::auto_moderation::AutoModerationRule
     AutoModerationRuleCreate = 140,
-    /// [`AutoModerationRule`] was updated.
+    /// [`AutoModerationRule`] has been updated.
     ///
     /// [`AutoModerationRule`]: crate::guild::auto_moderation::AutoModerationRule
     AutoModerationRuleUpdate = 141,
-    /// [`AutoModerationRule`] was deleted.
+    /// [`AutoModerationRule`] has been deleted.
     ///
     /// [`AutoModerationRule`]: crate::guild::auto_moderation::AutoModerationRule
     AutoModerationRuleDelete = 142,
-    /// Message was blocked by AutoMod according to a rule.
+    /// Message has been blocked by AutoMod according to a rule.
     AutoModerationBlockMessage = 143,
 }
